chore: update tui-sandbox from 3.0.0 to 4.0.0
