+++ integration-tests/cypress/e2e/example.cy.ts
@@ -1,6 +1,6 @@
 describe("testing", () => {
   it("can start neovim with a file", () => {
-    cy.visit("http://localhost:5173")
+    cy.visit("/")
     cy.startNeovim()
 
     // wait until text on the start screen is visible
