--- integration-tests/index.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/yazi-logo.png" />
-    <title>tsugit.nvim integration tests</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="module" src="./client/neovim-client.ts"></script>
-  </body>
-</html>
